Populate `packagetype` which is a required field
`packagetype` has been added to the model and is a required field. It
was not necessary when there was a single packagetype, sdist, but there
are now also wheels. We could probably assume that any package without a
package type is an sdist, but to be on the safe side we also check that
the filename has the correct file extension.